// webapp/src/api/model/characterSettings.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */
import type { InventorySizeEntry } from './inventorySizeEntry';
import type { BinaryExpressionTree } from './binaryExpressionTree';

export interface CharacterSettings {
  inventorySizes?: InventorySizeEntry[];
  maxPrimaryAttributeSum?: number;
  maxPrimaryAttributeValue?: number;
  minPrimaryAttributeValue?: number;
  talentPointFormula?: BinaryExpressionTree;
  tierFormula?: BinaryExpressionTree;
}